# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-commerce MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/shipments/{shipment_id}/advance": {
      "patch": {
        "tags": [
          "shipping"
        ],
        "summary": "Advance Shipment",
        "operationId": "advance_shipment_api_shipments__shipment_id__advance_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shipment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shipments/track/{code}": {
      "get": {
        "tags": [
          "shipping"
        ],
        "summary": "Track",
        "operationId": "track_api_shipments_track__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShipmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id"
          },
          "status": {
            "$ref": "#/components/schemas/ShipmentStatus"
          },
          "tracking_code": {
            "type": "string",
            "title": "Tracking Code"
          },
          "estimated_delivery": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Delivery"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "status",
          "tracking_code",
          "estimated_delivery",
          "updated_at"
        ],
        "title": "ShipmentOut"
      },
      "ShipmentStatus": {
        "type": "string",
        "enum": [
          "PREPARING",
          "IN_TRANSIT",
          "DELIVERED"
        ],
        "title": "ShipmentStatus"
      },
...
      "TrackingEvent": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ShipmentStatus"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "reached": {
            "type": "boolean",
            "title": "Reached"
          }
        },
        "type": "object",
        "required": [
          "status",
          "label",
          "reached"
        ],
        "title": "TrackingEvent"
      },
      "TrackingOut": {
        "properties": {
          "tracking_code": {
            "type": "string",
            "title": "Tracking Code"
          },
          "current_status": {
            "$ref": "#/components/schemas/ShipmentStatus"
          },
          "estimated_delivery": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Delivery"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TrackingEvent"
            },
            "type": "array",
            "title": "Timeline"
          }
        },
        "type": "object",
        "required": [
          "tracking_code",
          "current_status",
          "estimated_delivery",
          "timeline"
        ],
        "title": "TrackingOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}